+++ engine/func_profiler_simple_minmax.go
@@ -42,24 +42,24 @@ func NewSimpleMinMaxProfiler() *simpleMinMaxProfiler {
 }
 
 // Register a function for profiling
-func (p *simpleMinMaxProfiler) RegisterFunc(name string) int {
+func (p *simpleMinMaxProfiler) RegisterFunc(name string) uint16 {
 	id := p.base.RegisterFunc(name)
 	p.accum = append(p.accum, newSimpleMinMaxAccum())
 	return id
 }
 
 // Clear the data associated with a function
-func (p *simpleMinMaxProfiler) ClearData(id int) {
+func (p *simpleMinMaxProfiler) ClearData(id uint16) {
 	p.accum[id] = newSimpleMinMaxAccum()
 }
 
 // Start timing a function for simpleMinMaxProfiler
-func (p *simpleMinMaxProfiler) StartTimer(id int) {
+func (p *simpleMinMaxProfiler) StartTimer(id uint16) {
 	p.base.StartTimer(id)
 }
 
 // Stop timing a function for simpleMinMaxProfiler
-func (p *simpleMinMaxProfiler) EndTimer(id int) {
+func (p *simpleMinMaxProfiler) EndTimer(id uint16) {
 	ms := p.base.EndTimer(id)
 	accum := &p.accum[id]
 	accum.totalTime += ms
@@ -73,23 +73,23 @@ func (p *simpleMinMaxProfiler) EndTimer(id int) {
 }
 
 // Get the average runtime for a function
-func (p *simpleMinMaxProfiler) GetAvg(id int) float64 {
+func (p *simpleMinMaxProfiler) GetAvg(id uint16) float64 {
 	return (p.accum[id].totalTime /
 		float64(p.accum[id].invocations))
 }
 
 // Get the name of a given function
-func (p *simpleMinMaxProfiler) GetName(id int) string {
+func (p *simpleMinMaxProfiler) GetName(id uint16) string {
 	return p.base.GetName(id)
 }
 
 // Set the name of a given function
-func (p *simpleMinMaxProfiler) SetName(id int, name string) {
+func (p *simpleMinMaxProfiler) SetName(id uint16, name string) {
 	p.base.SetName(id, name)
 }
 
 // Return a string displaying the stats for a given function
-func (p *simpleMinMaxProfiler) GetSummaryString(id int) string {
+func (p *simpleMinMaxProfiler) GetSummaryString(id uint16) string {
 	return fmt.Sprintf("Summary for %s: {Avg: %.3f}",
 		p.base.GetName(id),
 		p.GetAvg(id))
