+++ engine/time_accumulator.go
@@ -10,14 +10,14 @@ package engine
 // private struct
 type TimeAccumulator struct {
 	// how much time has accumulated so far
-	accum int
+	accum uint16
 	// the periodicity of the time accumulator (used by both Tick() and
 	// Completion()
-	period int
+	period uint16
 }
 
 // Create a TimeAccumulator object with a given period
-func NewTimeAccumulator(period int) TimeAccumulator {
+func NewTimeAccumulator(period uint16) TimeAccumulator {
 	t := TimeAccumulator{}
 	t.accum = 0
 	t.period = period
@@ -30,7 +30,7 @@ func NewTimeAccumulator(period int) TimeAccumulator {
 // greater than the period (an odd situation, but important to note). `dt`
 // could be 100, and period could be 7, and we would still get a single true
 // value, even though really 14 complete periods had elapsed.
-func (t *TimeAccumulator) Tick(dt int) bool {
+func (t *TimeAccumulator) Tick(dt uint16) bool {
 	t.accum += dt
 	for t.accum >= t.period {
 		t.accum %= t.period
